Make writing StubModule a syntax error (#7624)
* Make writing StubModule a syntax error

From time to time we see crashes that occur because code generation
tooling generates lines like

    include Sorbet::Private::StubModule

which then causes downstream problems, because we assume that every time
a ConstantLit is `Symbols::StubModule`, we assume that
`resolutionScopes` was set.

But in fact, we only set `resolutionScopes` when constant resolution
fails. When a constant literal was written by hand and resolves
successfully to `Symbols::StubModule`, then we don't get
`resolutionScopes`, and accessing it later causes a Sorbet crash.

Let's make it impossible to write these special internal modules in
code, so that we can guarantee that if a constant literal is
`StubModule` that it was not written by a human and will necessarily
have come from constant resolution failing.

* Update exp files