Use different way to find LHS loc for && autocorrect
This broke after #7345, which we made in service of other autocorrects,
meaning that we couldn't assume `andAndLoc.beginPos()` pointed to the
right place (because it would technically include the enclosing parens).

Had we used the loc for the sake of wrapping the call to the `&&` or for
deleting the whole `&&`, we would have been able to do so. But we were
trying to find the start of the LHS operand to the `&&` by assuming that
the start of the `&&` operand is the start of the LHS inside it.