Acceptable test behavior changes
This change makes it so that a global constant (`class ::A; end`)
defined inside a package isn't treated as owned by that package (nor any
package).

It also makes it so that a package defined in the wrong namespace is
defined in the package whose namespace it would actually belong to (e.g.
`A::X = 1` belongs to `A` even if this definition happens in package
`A::B`).

And finally, this changes the location of some symbols in error
messages, but not in meaningful ways.