Remove the last use of getPackageForFile on symbols
This is the last place where we were using `sym.loc().file()` to get the
package that owns the symbol, rather than using symbol information to
lookup the package directly.

There are still a handful of `getPackageNameForFile` calls (which get
the `MangledName` for the package instead of the full `PackageInfo`),
and removing all the rest of these should mean that we can remove the
part of `VisibilityChecker` that calls `addLoc` to ensure that a
symbol's canonical loc matches the owning package.