Add some `const` in `SymbolFinder`
I'm working in this code, and it's helpful to know where the tree is
being modified. Basically the only place where we're currently modifying
the tree is to populate `ClassDef::ancestors`.

This doesn't matter much, but maybe it will make it more clear to
whoever needs to change this code in the future that they're the first
person to mutate that particular part of the tree in `namer`.