Remove `package:` argument from `non_forcing_is_a?`
This was a remnant of our earlier intention of making the package system
enforced at runtime. We ended up abandoning that approach, and you can
see because the implementation of `non_forcing_is_a?` says as much in a
comment (now deleted).

I'm going to codemod Stripe's codebase to no longer use the `package:`
argument (i.e., prepend it to the second argument) and then land this
static and runtime change.

I doubt that anyone outside of Stripe is using this.