More robust way of checking whether the packager is enabled (#7733)
* More robust way of checking whether the packager is enabled

Previously, we would use `packageDB().empty()` in a handful of places as
a proxy for whether `--stripe-packages` had been passed.

That only worked because the places that we needed to check this option
ran after the packager had already created packages--packager itself
used the command line option to determine whether to run the packager in
the first place.

As I start to move more of packager into namer, I need a reliable way to
determine whether `--stripe-packages` had been passed. Storing a boolean
explicitly is the easiest way to manager that.

Also, by setting it in the `setPackagerOptions` function, we ensure that
it's already going to be set in all the places that we'd need it to be
set, like in tests and in LSP, without having to go flip this bit on in
a bunch of places manually.

* Fix tests

* Reorder fields