Make findSymbols work in-place
This both allows us to remove a `fast_sort` at the end of findSymbols
and also sets this methd up to operate on `absl::Span`, which will make
implementing parts of the upcoming packager refactors more easily.

This could potentially adversely affect performance if this causes worse
cache locality, but I think that locality should be ~mostly fine as long
as the amount of work per file is approximately similar?

Another thing: before this change, we would not attempt to start merging
index results until the first thread had completed all its work, which
meant that merging the `allFoundDefinitions` was frequently the long
poll. Now, the merging happens as soon as the thread finishes, which
means it's keeping the main thread busier.