Unfortunately, we need the sort still
LSP runSlowPath was relying on namer sorting these files.
That meant that in the batch pipeline, which already sorts the input
earlier, namer was doing a redundant sort, but in LSP it was the only
sort.

We could ENFORCE this, except that we don't actually have the full list
sorted--`__package.rb` files are generally first, but only in the
pipeline, not LSP. So for now I'm just going to leave this new
`fast_sort` call here without an additional ENFORCE.