pipeline::resolve → pipeline::nameAndResolve
This changes the name one of the core pipeline functions, because we're
going to start to exercise more control around when the `name` portion
of it is called.

We could technically speaking leave it as a function and add a boolean
parameter indicating whether to call `name()` or not, but I figure
making new top-level functions is better.