Hoist `pipeline::name` to right after package
So now, we have the situation that every call to `pipeline::package` is
immediately followed by a call to `pipeline::name`.

In a future change, we'll basically merge most of `pipeline::package`
into `pipeline::name`, but in doing so, the package database will still
have been populated at the same point in the pipeline.