Add a test
I've decided to just check that the file is in the thing that's
returned, because I don't want to have this test changing every time
that someone adds a test above or below these assertions, messing with
the line number.

Testing that the `__FILE__` is in the `:caller_location` field should
bee good enough to test that the call to `caller_locations()` has picked
the right depth to pick this file, not some other file.