Reimplement what keep_for_ide did in DefLocSaver
It seems like we were using this `keep_for_ide` method only so that it
would get handled by our `postTransformConstantLit` callback in this
treewalk. But that's a lot of allocations to add to the tree when we
could just make `DefLocSaver` manually check the two places that have
extra constant lits.

This potentially removes a ton of allocations:

- All the `ConstantLit` / `UnresolvedConstantLit` that come from
  `deepCopy`'ing the constant in question.
- The `ConstantLit` for the `Sorbet::Private::Static`
- The `Send` for the `keep_for_ide` call

I don't think this removes the `<static-init>` methods if this causes
any `<static-init>` methods to be empty, but maybe that's an
optimization we could make.