Also make it work for Kernel.proc and Kernel.lambda
Making it work for `proc` and `lambda` requires knowing the type of
`self`, which means that the current strategy can't work.

We could make it work by unconditionally doing this rewrite for any
nullary method called `proc` or `lambda` called on `<self>` and then
make the intrinsic `<lambda T.let>` intrinsic dispatch to `proc` or
`lambda` if the method had been monkey patched, but that sounds like
more work than I want to do right now.