Don't dup documentationLocations when not aliased
The changes in 9c90faa74 accidentally made all documentation for
constants that were not constant aliases duplicated.

We should definitely include both the pre- and post-dealiased
documentation locations if the symbols are different, but if they're the
same, then it's just going to mean showing two copies of the
documentation.