Fix failing test
What's happening here is that Sorbet sees

    foo0 = Foo.
    foo1 = Foo.new(1, 2)

as the same as this:

    foo0 = Foo.foo1=(Foo.new(1, 2))

They both have this parse result:

    ❯ sorbet -p desugar-tree -e $'foo0 = Foo.\n foo1 = Foo.new(1, 2)\n' 2> /dev/null
    class <emptyTree><<C <root>>> < (::<todo sym>)
      foo0 = <emptyTree>::<C Foo>.foo1=(<emptyTree>::<C Foo>.new(1, 2))
    end

Which makes it hard for sorbet to distinguish this from a differen case,
where it likely wouldn't make sense to insert a snippet:

    x = self.already_has_args(1, 2)
    #                        ^ completion: ...

This change attempts to disambiguate that by checking whether the query
location is at the end of the line. This is simplistic (even a
whitespace character at the end of the line will break this heuristic),
but I think that eventually we should probably stop attempting to
generate snippets for method completion items because other language
clients don't do it, and there's a better LSP method to guide people
towards filling in arguments, namely signature helper.