Somehow ClassOrModule gets smaller?
It looks like absl made some changes to UnorderedMap which shaves off 8
bytes from `core::ClassOrModule`. It looks like that's because
UnorderedMap gets 8 bytes smaller. Somehow, it seems that this is the
only class inside Sorbet that we have a `CheckSize` annotation on that
uses `UnorderedMap`:

    ❯ rg -l -t h 'UnorderedMap' | xargs rg -l CheckSize | xargs rg -n UnorderedMap
    common/common.h
    24:template <class K, class V> using UnorderedMap = absl::flat_hash_map<K, V>;

    core/Types.h
    235:using IntrinsicMethodsDispatchMap = UnorderedMap<NameRef, const std::vector<NameRef>>;

    core/Symbols.h
    605:    UnorderedMap<NameRef, SymbolRef> members_;
    607:    UnorderedMap<NameRef, SymbolRef> &members() {
    610:    const UnorderedMap<NameRef, SymbolRef> &members() const {

    core/GlobalState.h
    338:    UnorderedMap<std::string, FileRef> fileRefByPath;

(`GlobalState.h` has a CheckSize but it's commented out, actually.)

So `UnorderedMap` gets smaller by 8 bytes and this `core::ClassOrModule`
benefits.

I tried to find the commit in abseil-cpp that introduced that change,
but couldn't find it after a few minutes of searching. In any case, a
welcome change.