Backport a patch from upstream to fix -Warray-parameter warning
I couldn't get this silenced by adding `--copt=-Wno-array-parameter`
because it looks like abseil builds this with a bunch of specific
warning flags, and those take precedence over ours.

The issue is fixed upstream, and when we upgrade we'll get this change
for free, but as mentioned in the previous commit we can't upgrade
abseil until after we upgrade clang.