Tweak TypeVar::_approximate to consult other bound
Before the changes in #6076 (dc0319bc98055a4a9f3676b1c48f721a8f8823c1),
the behavior of this function was "look at one bound, if there isn't a
bound for that, then use the other bound, and if there are no bounds,
use <top>."

After the changes in #6076, we started to **only** consult one of the
bounds, depending on polarity.

I'm curious to see whether we can improve the heuristic if we attempt to
smush those two heuristics together: always consult both bounds, but
choose which bound to consult first based on the polarity of the type
variable.

As dmitry wrote:

    // TODO: in many languages this method is a huge adhoc heuristic
    // let's see if we can keep it small

Maybe adding complexity here isn't great, but ¯\_(ツ)_/¯