Add failing test
Whether we report an error or not for both "Constants must have type
annotations with `T.let`" should not depend on file order..

In this test, we see that the type annotation for `A` comes in the
second file (after the file where `A` is defined without an annotation),
while the type annotation for `B` comes first (before the file where `B`
is defined without an annotation).

There should not be this sort of order dependency.