Add some debug log lines
We've seen a handful of crashes coming from here but I haven't been able
to come up with an example where this fails. Most of the crashes that I
see happen during file hashing, which suggests to me that the crash is a
pure function of the input file, which is a good sign that if we can
just get the file contents we'll be able to make a minimal repro and
then fix the issue.