Add patch to workaround Apple Silicon macOS build
The emsdk library uses a Python script to attempt to infer some
information about which `node` binary to try to run. That script runs on
the host directly, and uses `platform.machine()` to get the
architecture.

But earlier in the pipeline, the Bazel config uses information from
Bazel to figure out which nodejs external repository to fetch and
unpack (not Python).

Until we stop doing this weird Rosetta Bazel dance, we need to forcibly
have emscripten choose the version that the emsdk Bazel toolchain will
have fetched.

This also means that we won't be able to build our WebAssembly build on
x86_64 versions of macOS, but we don't do that during development nor in
CI (it still works to create WebAssembly builds on Apple Silicon macOS
and on Linux), but that restriction should be lifted once we can switch
away from running Bazel under Rosetta and drop this patch.