Remove the `sorbet.rename` VS Code command
This command is still available in the command palette as `>Rename
Symbol`.

We added this command because sometimes we want to register it as a
followup command from certain code actions. e.g. the LSP spec lets you
specify a command to run after making the WorkspaceEdit of a code
action. (The original context was the "Move method to new module" code
action, where we wanted to programmatically trigger a rename after the
method had been moved to a new module.)

That is: this command was only ever expected to be run programmatically,
not by users. (VS Code wouldn’t let us call put editor.action.rename
directly in the command field of the code action response for some
reason that I forget.)

If we ever want to bring this back, we should figure out how to declare
that this command can be used programmatically, but not show it in the
command palette. Or else, make it do something useful (i.e., not error
with an "undefined property" error when run from the command palette).