Also disable libtool for host
I'm not quite sure why this is required. Without this, I was seeing
problems because the list of arguments generated to be passed to
`llvm-ar` was completely wrong. Turns out, it's because it was trying to
craft a `libtool`-compatible argument list and pass that to llvm-ar, and
that's what this `--features=-libtool` thing is supposed to be turning
off in Bazel.

In 7.0.0, the `--features` flag only applies to the target, not the
host, and you need to specify `--host_features` to get the features in
both places if you want. This is controlled with the
`--incompatible_use_host_features` flag (which defaults to on in 7.0.0).