Don't hang if publishing fails
When `grep` isn't given a file to grep in, hangs without printing any
warnings. I think this is a pretty absurd default, and I even have a
blog post ranting about it:

→ <https://blog.jez.io/cli-tty/>

But this should fix the problem and make our publish.sh step not hang
indefinitely if the publish fails.

You can see an example of this hang happening in this build:

<https://buildkite.com/sorbet/sorbet/builds/31485#018fcad9-68c3-4c41-83b0-f3558eff2348/420-436>

Note the timestamps, it took ~2.5 hours before I realized that we had a
hanging build.