Move `--wait-for-dbg` into readOptions (#7978)
I wanted to be able to use `--wait-for-dbg` to debug some of the logic
that handles reading files recursively from the set of folders provided.
We read the files from disk as soon as we handle the `<files>` and
`--dir` argument in the `readOptions` function, whereas we handled the
`--wait-for-dbg` option after option parsing / reading files from disk
(in realmain).