Stop prefixing `./` in random places
It's especially annoying, because then we go and try to strip it out in
other places. That's annoying for two reasons:

- It's a waste adding `./` in some places then removing it in other
  places.
- It actually causes a problem when you try to use `--dir=subdir` to
  launch Sorbet from the root of a repository--some of the code deleted
  below (look for `substr`) would remove the `rootPath` from the
  `relativePath`, but we actually **can't** remove it (it only makes
  sense to remove when `rootPath` is `./`).

On top of that, we have hacks like `--remove-path-prefix` which attempt
to work around that everything is going to get prefixed with a path even
when you didn't want that.