Allow `# typed: __STDLIB_INTERNAL` to do what its designed for (#7979)
* Allow `# typed: __STDLIB_INTERNAL` to do what its designed for

The meaning of `# typed: __STDLIB_INTERNAL` is to permit overloaded
methods. Originally, that meant "in RBI files outside of those defined
in Sorbet's payload," thus the name `__STDLIB_INTERNAL`. But then in
7e4939fbdb3c86f477709a9e501b8d9dd5f1595d we relaxed the requirement that
overloads be defined in Sorbet's payload (allowing them in all RBI
files). I had tried to make it so that this strictness level would work
in non-RBI files to permit overloaded definitions but it seems like I
didn't test that.

It's fixed now

* Add some test assertions to files called `overloads_test.rb`