Remove the Hash case
This is causing a lot of problems because people are using these hashes
in tests where they assume they have the specified keys. I'd like to
think a little bit more before inferring shape types for hash literals,
so for now I'm just going to leave them as untyped and require
annotations in `# typed: strict` files.