Reapply "no-op: Factor isSingleton to take the whole type" (#8001)
* Reapply "no-op: Factor isSingleton to take the whole type"

This reverts commit 34af4ae581ad8a949f14f2ed83727d90971ac68f.

* Proper fix for the isSingleton code

What was happening is that we were including two knowledge tests:

- if the condition is true, then `value` has type `Unset`
- if the condition is true, then `value` has type `T.class_of(Unset)`

So then Sorbet would say that `value` had type `T.all(Unset,
T.class_of(Unset))` and then would collapse that to `T.noreturn`.

And because the `value` variable was live (used by a later block) it was
tracked as a block argument to the `when` block. Then when attempting to
compute it, Sorbet noticed that one of the block args was `T.noreturn`
and it reported an error that the whole block was unreachable.

The funny thing is that for the `module` case, this behavior was
correct: the block is unreachable because the `UnsetModule` type is
uninhatibed. Modules cannot be instantiated on their own, only via being
mixed into a class and then instantiating that class. But this module
was final, so that cannot happen. It's weird, in that we actually make
Sorbet _worse_ at detecting unreachable code, but the fact that it
reached the right conclusion in the `module` case was based on bad
reasoning, so we're arguably more correct here. We could potentially try
to add logic back for this case, but I don't really think it's worth it
(I would at least like to wait until someone shows me a program that
would benefit from such behavior).

* Update infer/environment.cc

Co-authored-by: Nathan Froyd <froydnj@gmail.com>

---------

Co-authored-by: Nathan Froyd <froydnj@gmail.com>