Inferring types for constants with arrays and constant literals (#7993)
* Allow inferring types for arrays of literals

* Add handling for hash

* TODO: consolidate TODOs

* Make it recursive

* Add support for certain constant lits

These constant lits are the ones that already have support in
type_syntax.cc. The `ClassOrModule` symbols are handled fully before
this phase of resolver. The `T::Enum` types are handled in this fixed
point loop, so it should be safe to depend on them here too.

* Fix crash in classes with not-yet-resolved type members

* Add a test

* Turns out we do need something like this

* Remove the Hash case

This is causing a lot of problems because people are using these hashes
in tests where they assume they have the specified keys. I'd like to
think a little bit more before inferring shape types for hash literals,
so for now I'm just going to leave them as untyped and require
annotations in `# typed: strict` files.

* Add a code snippet to the docs

* Fix tests