Some test changes
The `Enumerable#each` change is desirable: it's nice to let people
implement `Enumerable` with a `void`-typed block parameter.

The other two are because of how we treat conditional branches on
`void`. Before, we would do something like

```
cond : T.all(T.nilable(FalseClass), void_())
```

and this would be `T.noreturn` because `Sorbet::Private::Static::Void`
was not an ancestor of either `NilClass` or `FalseClass`. Now that it
gets treated like top, that doesn't happen (i.e., there is a shortcut in
`Types::all` where since `T.nilable(FalseClass)` is a subtype of
`void_()`, `Types::all` short-circuits to "just use the smaller type").