Treat `void_` as another top type in subtyping (#8012)
* Treat `void_` as another top type in subtyping

* Remove some `void_()` special cases

* Add a test

* Can't actually be as aggressive as I'd hoped

These are the tests that failed:

    //test:test_PosTests/testdata/compiler/abstract_static_methods
    //test:test_PosTests/testdata/compiler/absurd_raise
    //test:test_PosTests/testdata/compiler/nilable_block_param
    //test:test_PosTests/testdata/infer/branch_on_void
    //test:test_PosTests/testdata/infer/or_and_conditional_branch
    //test:test_PosTests/testdata/infer/pinning_type_parameter
    //test:test_PosTests/testdata/infer/square_bracket_err_loc
    //test:test_PosTests/testdata/infer/strong_block_return
    //test:test_PosTests/testdata/infer/strong_rescue
    //test:test_PosTests/testdata/infer/strong_splat
    //test:test_PosTests/testdata/infer/void_final
    //test:test_PosTests/testdata/lsp/fast_path/each
    //test:test_PosTests/testdata/infer/lub_between_self_params

* Some test changes

The `Enumerable#each` change is desirable: it's nice to let people
implement `Enumerable` with a `void`-typed block parameter.

The other two are because of how we treat conditional branches on
`void`. Before, we would do something like

```
cond : T.all(T.nilable(FalseClass), void_())
```

and this would be `T.noreturn` because `Sorbet::Private::Static::Void`
was not an ancestor of either `NilClass` or `FalseClass`. Now that it
gets treated like top, that doesn't happen (i.e., there is a shortcut in
`Types::all` where since `T.nilable(FalseClass)` is a subtype of
`void_()`, `Types::all` short-circuits to "just use the smaller type").