Infer lambda return types (#8011)
* Infer lambda return types in SolveConstraint

This doesn't rely on a large, hairy, tricky-to-get-right addition to
`guessOverload`, making it less brittle and faster at the expense of
maybe being less principled.

* Account for multiple BlockReturn

The previous implementation would be a "last `BlockReturn` node wins"
sort of arrangement, which meant that the inferred return type was wrong
any time the lambda returned early.

* Fix some tests

* Add a test

* Fix this test

* no-op: Reorder TypeArgumentRefs in SymbolRef.h

* Reimplement in Kernel#lambda intrinsic

This is a more principled approach than lurking `lambda`-related code
into environment.cc.

* Another test change

* Tests get better

This is the reason why the BlockReturn approach was unsavory. We had to
assume that if the type was `T.untyped`, that this was the first
BlockReturn, and start accumulating results after that. It meant we
could not distinguish between

- no values returned from the lambda, so it should be `T.noreturn`
- two values returned from the lambda, the first of which is
  `T.untyped`, so the whole lambda should be `T.untyped`

The generics-based approach solves this trivially.