Allow explicit nil to be serialized by enums (#7896)
This commit uses an `UNSET` constant as the default `serialized_val` on
`T::Enum`. This allows the runtime to understand the difference between
an explicit `nil` and an unset default when serializing enum values.
Sorbet already understands statically that a serialized value can be of
type `NilClass`, but the runtime was serializing as a string
representation based on the constant name instead.