Reapply all the good changes from #8050 (#8058)
* Add timers to GlobalState::sanityCheck

* ENFORCE -> ENFORCE_NO_TIMER in a lot of hot places

These are all hot for the sake of reading the initial GlobalState, and
are all ~trivial integer or boolean operations that should never take a
long time (vs something like an `absl::c_count` that maps a function
over a vector to compute a property).

* Use `constexpr` for `fuzz_mode` and `debug_mode`

In `--config=dbg` we don't necessarily get optimizations that would
remove this check. I want to make sure it gets removed and that's easy
with `constexpr`.

    fastmod 'if \(fuzz_mode\)' 'if constexpr (fuzz_mode)'
    fastmod 'if \((!?)debug_mode\)' 'if constexpr (${1}debug_mode)'
    tools/scripts/format_cxx.sh

* Remove some counters that are not that useful

* Factor out sanityCheckTableSizes function

* Factor out sanityCheckNames function