Ignore magic method <required-ancestor-lin> during documentSymbols
By design requires_ancestor keeps track of its usages by storing the loc
of where it was defined. If requires_ancestor defined on Foo is used in
thousands of files thorugh mixin of Foo then there would be thousands of
`<required-ancestor-lin>` methods that point to the loc of Foo. This
would mean documentSymbols will try to return all of those methods when
`foo.rb` is opened which will crash the LSP.