Refactor requires_ancestor magic method
Before we were storing relevant information in both arguments and return
type. This PR cleans it up a bit and only uses arguments.

Each argument to the magic <required-ancestor-lin> or
<required-ancestor> methods is a tuple where first element is the
ancestor being required and the second is the constant that required
it. Each time a module which defines an ancestor is mixed in we create a
new argument.