Change optparse.rbi T::Class overload
I'm not sure why, but when I try to take advantage of this overload, it
doesn't work at runtime:

    ❯ cat foo.rb
    # typed: true
    require "optparse"
    require "optparse/time"

    OptionParser
      .new do |parser|
        parser.on(Time, "--time [TIME]") do |time|
          p(time)
        end
      end
      .parse!

    ❯ ruby foo.rb -t 9:30
    foo.rb:7:in `block in <main>': argument type given twice: NilClass (ArgumentError)
            from foo.rb:6:in `new'
            from foo.rb:6:in `<main>'

(It's not clear from that error, but the error comes from inside the
optparse.rb gem, which does some filtering of the exception backtrace to
hide the optparse.rb frames.)

I'm not sure if that's because something changed and I'm not sure what,
or if it never worked and I didn't actually test the runtime when I
wrote the original change.

In any case, it seems like optparse expects the type argument to come
after any short or long option names. We can work around that with some
more overloads.