website: Add script to make new doc (#8132)
I think that it's important to have low activation energy to create new
docs or else people won't do it. This script creates you a file,
populates it with the right boilerplate, and edits the sidebars.json
file to mention your new doc. It will even open the files for you if
`$EDITOR` is set.

I was trying to figure out the best way to preview the site but I
couldn't figure out something good. For now I think people will want to
run the command to preview the site in another terminal.