Reduce the ForwardedRestArg case to the Splat case for sends
We already have support for splats in method call argument lists and in
arrays. But previously, we only had support for ForwardedRestArg in
method call argument lists, not in arrays.

There's no reason why it has to be that way, and in fact using it inside
an array was raising an uncaught exception. We can delete a lot of code
and fix the bug by reducing the ForwardedRestArg to the Splat case.

(Note that this works because the splat case in method call arguments
**already** works by reducing argument lists to the array case.)