Remove rubyRegionId
This was added in service of the Sorbet Compiler, which is now deleted
(#7849).

It had the unfortunate consequence of bloating the CFG in a way that
only mattered for compiling code, not for type checking code.

We should be able to put it back to what it was and save a little
memory (and have less code to read and wonder about its meaning).