Use a `repository_rule` to get the RUBY_ROOT env var
Up to this point, in order to build Prism, we've needed to manually pass
the path to the correct RUBY version to the build process because Bazel
does not take the user's environment into account and will just use
the system's default Ruby version (which is often very old).

This is a workaround for our workaround, where we use a Bazel
`repository_rule` to get the `RUBY_ROOT` environment variable,
write it to a file within the Bazel file system, and then pull
the value from that file while building Prism.