Extract ensure/rescue clause from method body if no begin block is present
If the method body is a `PM_BEGIN_NODE` instead of a `PM_STATEMENTS_NODE`,
it means the method definition doesn't have an explicit begin block.
If that's the case, we need to check if the body contains an ensure or
rescue clause, and if so, we need to elevate that node to the top level
of the method definition, without the `Kwbegin` node to match the behavior
of Sorbet's legacy parser.