Replace unused constructor overload
On this branch we've replaced both usages of the

    DispatchResult(TypePtr, DispatchComponent)

overload, so we don't need it. Also, I want to add another constructor
overload to make it easier to implement the other cases.

The `explicit` makes it so that you can't implicitly convert from
`DispatchComponent` to `DispatchResult`, which would likely mean there's
a logic error somewhere.