callWithSplat: Overwrite the DispatchResult
No need to be clever here: just overwrite it, and behave as if the
dispatch happened to something else.

This works because there's nothing that could have reported an error
earlier, before the call to the intrinsic kicked in: the method itself
takes untyped arguments, so there will never be anything important on
`res.main` worth saving.