Change the origin of non-synthetic assignments
Sorbet will make `cfg::Ident` instructions in response to both
`ast::Assign` nodes with a source of a variable (`x = y`) and also
`ast::Local` nodes (`y`).

Sorbet previously said that the "origin" of `x` above was the same as
the origin of `y`. But when `x` is not a synthetic temporary, and was
instead some assignment that the user actually wrote, we should treat
the origin of `x` as the instruction itself in that case.