Track parse errors in addition to index errors
In some cases (like whether or not to recommend `typed` sigils) we only
want to gate that behavior on whether the file parses successfully. This
suggests that we should track the presence of parse errors in addition
to indexing errors, with the relationship looking like

`parse errors => indexing errors`

i.e., it's fine to have indexing errors without parse errors, but parse
errors indicate the presence of indexing errors.