`nullptr` check during `ENFORCE` in `downcast`
In our `downcast` helper, we check that the node argument is actually
a Prism node. However, doing this requires us to check the `type` field,
which will fail if the node is a `nullptr`. By doing a `nullptr`
check in the `ENFORCE`, we guarantee we won't crash by checking for
a field on a `nullptr`.

Co-authored-by: Alex Momchilov <alexander.momchilov@shopify.com>