Also send Timers' tags to the --web-trace-file
It's not clear to me why we weren't doing this before.

We don't programmatically compute or ingest `--web-trace-file` output at
Stripe--it's only for manual debugging, so this should not make any
critical workflows slower.

Importantly, it's not actively computing or storing any new information:
this information was already being stored on all these timers for the
sake of statsd tagging.