Two fixes to "convert to singleton method" action (#8288)
* Compute convert_to_singleton_method with two edits

The old mechanism for crafting the edit worked by completely
reconstructing all the constituent parts, which had the nice property
that there was one edit for one code action.

The problem is that Sorbet doesn't currently track a location for "all
the stuff inside the parens," only the location of the first arg and the
location of the last arg. Normally those are the same but they're not
the same in the case when either the first or the last arg is wrapped in
parentheses:

    foo((A), (Z))
         ^^^^^^

Here, the range from the beginning of the first arg to the ending of the
last arg does not include the parentheses. A previous change of mine
attempted to make that be the case
(7eb2d87f56aa543e6e6a139ec59c0ee9a6e3413c) but that ended up introducing
a crash in resolver, and also made anything that actually did just want
to deal with the what was inside the parens more cumbersome.

We don't actually need a location for all the arguments, we just need to
know where the first one is. If there are already arguments being
passed, that's easy. If there aren't, it's a little tricky, because we
have no location that tracks where an empty `()` in a method call is:

    foo()
       ^^ nothing tracks this location

So to avoid editing the file like `A.foo(self)()`, we have to attempt to
detect that case on our own. We can assume that the empty `()` will be
immediately after the `funLoc` if it's there at all, because `foo ()`
means "pass one empty InsSeq argument to foo" instead of "call foo with
zero arguments."

* Add failing test

* Fix failing test

* Fix sanitizer failure

`receiverSource` was a `string_view`, which means that the string
computed by `fmt::format` would not necessarily live as long as
`receiverSource`.

* Change the CheckSize assertion