Don't duplicate conditional branches when outerLoops differs (#8291)
* Add a regression from 8283

* Fix a typo in cfg simplification

Update the backEdges in a way that's consistent with the else
implementation below.

* Maintain unconditional branches when short-cutting

This keeps the CFG sound, otherwise we may update half of an
unconditional branch and then fail the sanity check. Usually this isn't
a problem because block merging happens before short-cutting, so we'll
not usually attempt short-cutting in an unconditional branch, but
debugging single transformations makes this change necessary.

* We shouldn't duplicate conditional branches that change loop level

Duplicating loop headers will throw off flow-sensitive typing, and break
the topological order. As a result, we will see variables before we know
anything about them, and often incorrectly infer that they are nil at
that point.

* Update the test for cfg changes