Move vector allocation out of a loop over all files
We know the maximum size of the intersection vector, and we can use that
knowledge to amortize the allocation across the loop over all files in
the project. This will avoid thrashing the heap if there is a popular
name in the project, and should mean that the only additional
allocations come from emplacing file refs in `extraFiles`.