Stop surfacing Prism's EOF error
It usually appears when there's an unclosed statements/assignments, where
the root error will also be surfaced. So it's not exactly important to
surface this error.

But more importantly, because Prism puts this error at the very last line
of the file when there's a missing `end`, it can't be represented in
Sorbet's error comments.