Translator should not treat `~[Integer]` as a method call
Sorbet's legacy parser treats `~[Integer]` and `-[Integer]` as integer literals,
but Prism treats `~[Integer]` as a method call.

(Technically, both of them are method calls.)

So in the translator, we need to give `~` special handling when its receiver is an integer.
We don't need the same for Float or Numeric because only Integer has the `~` method.