Unconditionally call `setHasIndexErrors`
Without this, we saw an error in the `//test/cli:test_errors-skip-cache`
test where the duplicate type member error (coming from rewriter) would
not be reported on the fast path.

(The "Duplicate type member" error uses the error code 4011 as a
carry-over from when namer reported that error. It's now reported
syntactically in rewriter, but we kept the old error code so as to not
invalidate anyone who might have been silencing it.)

It doesn't seem like we need to check the error code at all. I think
that this guard might have been a carry over from an earlier
implementation of `beginIndexerError`, when were planning to just use
`beginError` itself, instead of making a separate method.