Improve handling of `uniqueCounter` in Prism translator
`uniqueCounter` is a counter that is used to give unique ids to anonymous arguments
while parsing them in Sorbet. Thus far, we have coordinated the `uniqueCounter` between
multiple Translators by passing the value back and forth, but this approach is brittle.

With this new approach, we can maintain the same `uniqueCounter` between multiple
translators:

1. The parent translator maintains the source-of-truth counter in a variable called
  `uniqueCounterStorage`
2. It also maintains a pointer to that value called `uniqueCounter`
3. Child translators are created with a dummy value in `uniqueCounterStorage` that
  will never be used, as well as the `uniqueCounter` pointer that points to their
  parent translator's storage
4. Incrementing the `uniqueCounter` happens via the pointer, which will always
  point to the parent's `uniqueCounter` value

Co-authored-by: Alexander Momchilov <alexander.momchilov@shopify.com>