Creating abstract wrapper methods should not trigger `method_added` hook
The previous commit that changed how abstract method wrappers are created missed a `without_method_added` block. This caused the `method_added` hook to be triggered when creating abstract method wrappers, which is not the intended behavior.

Moreover, since the new method definitions could trigger method redefinition warnings, it is better to also wrap the method definition in a `without_warnings` block.