use raw pointers in `DispatchArgs` for efficiency/clarity's sake (#8422)
So, imagine you have:

```c++
struct SendAndBlockLink {
  ...
};

struct Send {
  std::shared_ptr<SendAndBlockLink> link;
  ...
};

struct DispatchArgs {
  const std::shared_ptr<const SendAndBlockLink> &block;
  ...
};

// Somewhere in your code...
DispatchArgs args{send->link, ...}
```

what happens on that last line?

The obvious to the point of C++ 101 intent is that we take a reference to `send->link`.  Straightforward, right?  And that would be what happens if `block` had been of type `const std::shared_ptr<SendAndBlockLink>&`.  But since the `shared_ptr` is storing a pointer to a _`const`_ object, the types for the kind of references we need are different.  So there needs to be a temporary `std::shared_ptr<const SendAndBlockLink>` constructed from a `std::shared_ptr<SendAndBlockLink>` and then we can take a reference to that temporary object -- and then destroy the temporary `shared_ptr` when we are done.  (This also happens completely invisibly to the user, as you can see.)

This is obviously unnecessary overhead, which is bad.  That is not the actual problem here (though we'd like to avoid it).

Welcome to the fun of language arcana.  How long does that temporary object last?  We really want it to last as long as `args` is live, rather than just be live for the construction of `args`.  If the lifetime of the temporary object ended once `args` was constructed -- but before `args` was used -- any use of `args.block` would be the equivalent of use-after-free.

C++ recognizes that this sort of thing needs to happen automatically in some cases; it's called "lifetime extension".  The lifetime of the temporary object we construct is magically declared to be longer than `args` itself.

Due to the rules of the language, the lifetime extension is automatically applied only when we initialize `DispatchArgs` via brace-initialization.  If we were to, say, give `DispatchArgs` a proper constructor, lifetime extension would no longer apply and [we would get a litany of confusing ASan errors about stack-use-after-scope](https://buildkite.com/sorbet/sorbet/builds/33874).

IIRC, we need to give (the Sorbet version of) `DispatchArgs` a proper constructor to make certain things work in C++20 -- or at least we did, once upon a time, maybe that's no longer necessary -- and I think it'd be nice to a) eliminate the performance issue involved in constructing an entirely new `shared_ptr` object and b) eliminate this completely arcane behavior.  Both of these considerations apply independently of whether we still need the constructor IMHO.

Hence this PR.  I know we don't love raw pointers in Sorbet, but I think the alternatives are worse: things like:

```c++
const auto &block = *reinterpret_cast<std::shared_ptr<const core::SendAndBlockLink>*>(std::addressof(send.link));
```

when initializing call dispatch in inference ([which does work](https://buildkite.com/sorbet/sorbet/builds/33880)!  but that build also shows that we'd need to apply similar behavior in other contexts too.  I think I prefer the raw pointers, all things considered).  It's also nice that we don't have to heap allocate things for splat calls in `calls.cc` anymore.