hasKwargs → hasKwparams
args are at the call site, params are at the definition site.
This method plays fast-and-loose with that convention, especially
because the list of a method's parameters are stored on a field called
`arguments` on the symbol. But we already use `ait` and `pit` for the
argument and parameter iterator in this method, and this code is hard
enough to read as it is.

Eventually, we should even change `core::Method::arguments` to be called
`parameters` and `ArgInfo` to `ParameterInfo` but that's a refactor for
another day.