Avoid updating the file table on the slow path (#8569)
* Avoid updating the typechecker's file table on a slow path

The slow path is already replacing the global state with a copy of the
indexer's global state, and all files that we see as updates should
already be inserted in the indexer's global state. As a result, we don't
need to manipulate the file table at all, and can simplify the updates
we perform.

* Return copied trees from `copyIndexed` instead of using an out parameter