Still have to treat it like their underlying is Object
There are 41 type errors in Stripe's codebase from this and then a
handful more from typechecking sorbet-runtime, and I don't actually want
to fight this battle today: I just want to deprecate this
`expliainTypeMismatch` method.

Most of the violations come from:

- `==`
- `!=`
- `puts`

So it's reasonable to assume that we could do this migration in the
future and stop somtimes treating metatypes like `Object` but that's a
battle for another day.